Bug 423105 - Improve parent pom
Add prereq section to define that maven 3.0.5 is the min required
version due to cbi-plugins requiring it. This will make it obvious from
the start if one tries to build with old Maven.
Also update maven-resources-version to version 2.6 which supports
maven.test.skip parameter and will skip resources calls to shave the
build time a bit.

Change-Id: I3a8b3128779e64f4022e116d5ea59368b198e0c4
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>