Bug 500233 - Build IDs 1 minute too early
Improved fix truncating epoch seconds to nearest 5 minutes. 

Modified previous improvement to remove "warning" and simply 
echo another "DEBUG" statement that "RAWDATE was already set".

I observed in a test run that the source file is "included" several
times, so it is misleading to declare it a "warning" if it is found
already set.