[releng] this 'publish2.xml' file is not (yet) used.
In production builds, that is. It was the "nearly finished" version of
publish.xml that would do a better job of publishing performance unit
tests. (Along with other changes). 

I am not sure if it will ever be used, but it had a few obvious ant
syntax errors that were showing up as "red X's" so best to fix.