Bug 533905 - Fix and simplify Oomph setup for JDT
Ensure that an API baseline is set; it should be Oxygen now.

Set eclipse.target.platform to eclipse.target.platform.latest which is
maintained in the project catalog; it is Photon now (and don't set it in
any of the streams).

Set eclipse.api.baseline.target.platform to
eclipse.api.baseline.target.platform.latest.released which is maintained
in the project catalog; it's is Oxygen now.

Include a task to resolve JDT's API baseline and set it; this task is
composeable with the setups of other projects, producing a common
baseline.

Simplify the targlet tasks, e.g., always use
http://download.eclipse.org/cbi/updates/license regardless of the
current target platform choice.

Change-Id: I10ac3a686f7cfb57084be00c26aa157593ca78ea
Signed-off-by: Ed Merks <ed.merks@gmail.com>