Bug 540222 - All test projects should dump stack traces on timeout
When running tests with org.eclipse.test.EclipseTestRunner during
integration builds, stack traces and screenshots are provided if the
build times out (e.g. due to a hang in the tests). When running test
projects with maven however (e.g. when jdt.debug.tests run in a Gerrit
job), no such stack traces or screenshots are present. Debugging
sporadic hangs in Gerrit jobs is therefore very difficult.

This change is a part of moving the dump-on-timeout functionality to a
location which is accessible to test applications, namely in
org.eclipse.e4.ui.workbench3 (along with
org.eclipse.ui.testing.TestableObject).

Change-Id: I301883c95a7fe68ae8102401b6ec6cb4e9634e06
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>