Bug 563523 - Set a project name to allow to distinguish projects
There are two projects that have identical artifactids what makes it
impossible to differ between the projects in the build log.
I am pretty sure that changing this can have a lot of impact on other
project consuming these projects. Although changing it would be the
right thing for the moment it is imho an improvement to just fill out
the project name.

Change-Id: I650aa04072ff43a2800badfca6ee2a9f4e67dbf4
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>