Bug 567827 - Stop building and publishing releng.tools zip
It's in the p2 repo and is better installed that way so all deps are
autoresolved and installed too.
Not to mention it's one step less to install from the p2 repo directly
as when using the zip you have to download and set it up as p2 repo
before installing from it.

Change-Id: I64c3f9022ae82d700df43e256f06478f70103ccb
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>