Always check for API analysis when profile is on #174
In some cases, the skipAPIAnalysis=true property can be exported from
Ant into the Maven session or some parent (instead of current project
properties). This might be related to pom-less generating a static value
for this property.
The `skipAPIAnalysis` is useless here anyway because people who want a
flag to skip API Analysis can explicitly exclude the profile already.