SDK4Mvn: Remove default patterns
Default patten mappings have high probability of being incorrect for a
random dependency, even more if the artifact initially comes from Maven.
So instead of adding some rule that covers that case wrongly, we add
nothing an may have the build explicitly failing and requesting a
dedicated rule if none of existing ones matches. The failure becomes a
hint that this file requires some manual update as a reaction of a
dependency change.
Also specify some explicit rules for some artifacts.