Fix broken link in Performance Table
xml links in https://download.eclipse.org/eclipse/downloads/drops4/I20220802-1800/performance/performance.php are broken. 
The php file includes an html file that's one directory up from it. The html is written to work from that location and it used to work when pulled into the php file but for some reason now the paths need to work from the php files directory. This should fix it.