Simplify the API removals document
Removal the manual created TOC which is not complete and therefore
misleading. Also updating this TOC is painful and it is not checked
during the release validation process (otherwise I would not miss
entries).

Also moves the deleted API to the end of the document, I assume the
people are not interested on what is to come, than the past deletions.
Also we have now only two links, one to the planned deletion, on the to
actual deletion.

See also https://github.com/eclipse-platform/.github/issues/22 for an
even simpler suggestion for API removals.