Temporarily remove bouncycastle refs from eclipserun
Those bundles currently are not available in the I-Build repo that's
used by eclipserun to get content. So we remove them until they're
available.
In the meantime, build should work better and not block, but there will
probably be Javadoc issues, that will be resolved when re-introducing
the bundles here.