Add BND lib to the target
m2eclipse already uses BND for a while to generate missing manifest for
maven dependencies and PDE can (re-)use BND as well for different tasks.

This also moves the bnd annotations to the new BND section and adds
org.osgi.service.repository API bundle as it is a dependency of BND lib.