Make SVG files smaller (#112)
* Make SVG files smaller

SVGO is a cool to to minify SVG files (by e.g. removing invisible
objects).
Let's use this to squeeze some bytes out of all the SVGs used in
documentations

Fixes: #111
Signed-off-by: Matthias Becker <ma.becker@sap.com>

* Make SVG files in org.eclipse.platform.doc.isv smaller

SVGO is a cool to to minify SVG files (by e.g. removing invisible
objects).
Let's use this to squeeze some bytes out of all the SVGs used in
documentations

Fixes: #111
Signed-off-by: Matthias Becker <ma.becker@sap.com>

* Make SVG files in org.eclipse.pde.doc.user smaller

SVGO is a cool to to minify SVG files (by e.g. removing invisible
objects).
Let's use this to squeeze some bytes out of all the SVGs used in
documentations

Fixes: #111
Signed-off-by: Matthias Becker <ma.becker@sap.com>

* Make SVG files in org.eclipse.jdt.doc.user smaller

SVGO is a cool to to minify SVG files (by e.g. removing invisible
objects).
Let's use this to squeeze some bytes out of all the SVGs used in
documentations

Fixes: #111
Signed-off-by: Matthias Becker <ma.becker@sap.com>

Signed-off-by: Matthias Becker <ma.becker@sap.com>