Unify the platform workflow by defining reusable workflows
Currently there is a quite inconsitent mixture of ci workflows, some are
outdated or use different aproaches and configs, some have no workflow
defined at all.

This defines (and documents) a set of reusable workflows that can be
adjusted / updated at a central place and then reused in the platform
repositories.