Move common build arguments to maven.config and remove unnecessary
The property 'project.build.sourceEncoding' is already defined in the
eclipse-platform-parent/pom.xml used as parent for all modules, thus it
is not necessary to explicitly add it in the Jenkinsfile.

Moving the definition of the 'aggregatorBuild' property to the
maven.config simplifies local builds because then the property does not
have to be declared.