Don't include timestamp in javadoc
Currently there is a (hidden) timestamp generated in the javadoc what is
not very useful for our use case where we want to replace artifacts by
the baseline version what means we have changes on each build.

This disable the generation of timestamps for javadoc to get more
reproducible builds.