Prevent two PRs for same update from dependabot
Example is
https://github.com/eclipse-platform/eclipse.platform.releng.aggregator/pull/1629
and https://github.com/eclipse-platform/eclipse.platform.releng.aggregator/pull/1627

The config on "/" is enough as eclipse-platform-parent is a module of
the top pom.
This is leftover from the past as initially the repo was mostly pointing
to other subrepos and only eclipse-platform-parent was considered for
updates. When config for / was added the one for eclipse-platform-parent
should have been removed.