Declare required jdk and maven as top-level pipeline tools in I-builds
Since https://github.com/eclipse-platform/eclipse.platform.swt/pull/973
the I-build (like the SWT build) requires a java-17 (or later) jdk on
the PATH system environment-variable in order to launch a jdk>=17 with
the 'java' command from the CLI. Just setting the JAVA_HOME variable to
a path pointing to a jdk-17 is not sufficient since the the java
executable is usually located in '$JAVA_HOME/bin'. And while the maven
launch script searches in '$JAVA_HOME/bin' for a java executable to run
the build itself, a simple java command on the CLI isn't that smart.
Previously running the 'java' command would just run the default default
java version on the Jenkins executor.

In general this simplifies the build-configuration because the JDK to
use is only defined once and used everywhere.

Fixes https://github.com/eclipse-platform/eclipse.platform.releng.aggregator/issues/1724