Do not disable the default metadata generation
Currently the eclipse-sign goal disables the default meta data
generation of a project (as it needs to update the metadata afterwards)
but this seem to fail under some circumstances with quite obscure
errors.

This removes the disabling of the default-metadata what might be a tiny
little slower but seem not to cause any issues in reward.