Print stacktrace on failure of maven builds
Currently stack trace are suppressed on maven build problems, as those
are often the only way to understand unexpected build failures this now
enables them by default. The drawback is that the output is a bit harder
to read, but as we already have parsers for compile and test logs that
present this to the user in a more comfortable way it seems a good
compromise.