[Build] Define cron-trigger for DockerImagesBuild in Job definition
The triggers defined in a Jenkins-pipeline are only applied to the
Job-configuration after the first execution of the pipeline.
This has the consequence that after each execution of the seed job the
configuration of the 'DockerImagesBuild' is reset to have no triggers
and consequently it would require a manual execution of that job to
restore the cron-trigger each execution of the seed job.