[I-Build-Test] Remove CPU-arch as criterion for testPlatform property
All usages depending on the 'testPlatform' variable don't distinguish
between architectures.
In case it is attempted to run the tests on a unsupported arch, checking
the architecture too produces an explicit error message instead of just
having a download that fails because the corresponding artifacts are not
available. But those that run I-builds locally or configure a new
platform are usually well aware if the executing architecture is
supported or not.
And testing for the arch just makes it more complicated to run tests on
new platforms.