Add manual trigger for 'updateRelease'-workflow and use project-bot
Adding a manual trigger allows to re-run that job manually with the
latest changes applied to the referenced workflow from the
releng.aggregator repository.

And since we have a eclipse bot specifically for each project we can use
that bot as author of these automated changes instead of the generic GH
bot.