Define a single property for the releaseVersion and use in property
Actually the both properties 'releaseNumberSDK' and
'releaseNumberPlatform' are always the same and even if the comment
mentions they might be different at some time it is better to unify them
until then.

Because of this there is now a new 'releaseVersion' defined as a single
property.
Use placeholder to define the integration build repository

Currently we use fixed strings that regularly needs to be updated, but
the contents could actually be computed automatically.

This replace fixed strings with using properties making it needless to
change them all the time.