Rename gpg pashprase env variable for maven
Fixes warnings :
```
12:37:46  [WARNING]  Parameter 'passphrase' (user property
'gpg.passphrase') is deprecated: Do not use this configuration, it may
leak sensitive information. Rely on gpg-agent or env variables instead.
12:37:46  [WARNING]
12:37:46  [WARNING] W A R N I N G
12:37:46  [WARNING]
12:37:46  [WARNING] Do not store passphrase in any file (disk or SCM
repository),
12:37:46  [WARNING] instead rely on GnuPG agent or provide passphrase in
12:37:46  [WARNING] MAVEN_GPG_PASSPHRASE environment variable for batch
mode.
12:37:46  [WARNING]
12:37:46  [WARNING] Sensitive content loaded from Mojo configuration
```

from I-builds.
Rename is being done only for this step to get rid of the warning as for
the other steps it's not a maven env variable (thus shouldn't be named
like that) and will require changing multiple scripts.