[Build] Simplify bootstrapping of I-build tests and fetch files from git
Instead of fetching the test-configuration files from eclipse.downloads
and push it there in each I-build, simply get the test-configuration
files directly from the eclipse.platform.releng.aggregator git
repository, using partial-clones for improved performance:
- https://git-scm.com/docs/partial-clone
- https://github.blog/open-source/git/get-up-to-speed-with-partial-clone-and-shallow-clone/

This avoids the need to pack and upload the test-configuration zip-file
and also simplifies fetching the test-configuration for other branches,
like maintenance branches or to test changes.

This also removes from the getEBuilder.xml ant script the possibility to
download the test-configuration zip file and to use a local zip.

Furthermore this removes from the I-/Y-builds the creation and
publication of the test-configuration zip-file.