[Build] Remove getEBuilderxml obsolete parameters and simplify its path
Remove the definitions of the properties 'EBUILDER_HASH' and
'downloadURL', which are not necessary anymore to be passed to the
'getEBuilder' ant-script. And remove the 'eclipseStream' property that
now can be computed by the 'getEBuilder' ant-script.
Also rename the folder containing the getEBuilder.xml from to just
'bootstrap' since Hudson is not used anymore and simplify the path to
the getEBuilderxml at the download server.