[Build] Define complete configuration of StartSmokeTests job via Job-DSL
This way the complete configuration of the 'Start-smoke-tests' job is
under version control.
The only change is that this moves the 'Start-smoke-tests' job from the
root of the Jenkins instance into the 'SmokeTests' folder, where all
dedicated smoke-tests are located too.
This simplifies and cleans-up the structure of the Releng-Jenkins.