[Build] Define JDK-URLs within the StartSmokeTests Jenkinsfile
and only use the latest Temurin JDK of the corresponding (feature)
release in all smoke tests.

Defining the URLs within the Jenkins file has the advantage that changes
are immediately visible upon the next execution without the need to run
the seed-job again and without the one-job-delay that happens if
parameters change.