[Build] Unify declarations of list of executed I-/Y-build tests
This reduces the number of places to adapt in case the I-/Y-build
test-configurations are changed and in general reduces the difference
between I-builds and Y-builds.
Furthermore it reduces duplication introduced to handle the small
difference between I- and Y-builds.

Part of https://github.com/eclipse-platform/eclipse.platform.releng.aggregator/issues/2625