[Build] Unify declaration of build-properties for I-, Y- and P-builds
Instead of duplicating the 'buildproperties.txt' template for each build
type, keep only variables with common values in the general
'buildproperties.txt' and define differing variables in the Jenkins
pipeline and add it dynamically to the generated/derived
'buildproperties.shsource/properties'.

Remove unused 'TESTED_BUILD_TYPE' variable.
Update 'API_PREV_REF_LABEL' for Y- and P-build to be in sync with the
value defined for I-builds.

Part of
https://github.com/eclipse-platform/eclipse.platform.releng.aggregator/issues/2625