[Build] Fix export of 'TEST_CONFIGURATIONS_EXPECTED' in SSH execution
Because in the 'Collect Result' jobs the publish.xml ANT script is
executed remotely on the download-server via SSH the environment
variables defined in the shell on the build-server are not available.
As a workaround, define the 'testsConfigExpected' property explicitly as
command-line argument.

Also rename 'TESTS_CONFIGURATIONS_EXPECTED' to
'TEST_CONFIGURATIONS_EXPECTED' to avoid a double-plural.

Follow-up to
https://github.com/eclipse-platform/eclipse.platform.releng.aggregator/pull/2693