[Build] Update, clean-up and use job for Maven publication
Currently the publication of the Maven artifacts of Eclipse happens in
five jobs manually/directly defined in the 'Releng' Jenkins instance:
1. CBIaggregator
2. RepositoryValidator
3. PublishPlatformToMaven, PublishJDTtoMaven and PublishPDEToMaven

Corresponding job-definitions for these jobs already exist in this SCM
(except for the repository-validation), but the jobs generated from them
(in the 'Releng' folder instead of the root) are not used.

This change replaces the existing (unused) multiple job-definitions by a
single Jenkins pipeline that contains all steps and performs the final
publication for all covered projects in a single matrix.
The new pipeline is updated to incorporate all the changes made in the
meantime to the manually maintained publication jobs.

The I-build job is changed to use the new publication job.

With respect to the existing, manually maintained publication job, the
new job is adjusted in the following aspects:

- Rename job to just 'PublishToMaven'. The CBI part by itself is
irrelevant and a tool in the publication process
- Remove documentation about the jobs involved and scripts called from
the job description as it's an implementation detail that can be looked
up in the pipeline if of interest.
- Don't require the 'projects-storage.eclipse.org-bot-ssh' SSH agent for
the aggregation step.
- Only run publication on explicit invocation (from I-builds). Remove
the timer-trigger: Without changes a (re-)publication is not necessary.
- Make '-snapshot' the default value of the 'snapshotOrRelease'
parameter instead of '-release'.