[Build] Clean-up, inline and update maven deployment configuration
Move the configuration of the maven-gpg-plugin into the
eclipse-platform-parent/pom.xml and update it to the latest version.

Remove the configuration of the 'nexus-staging-maven-plugin' as it is
effectively unused because the 'gpg:sign-and-deploy-file' deploys
directly without using the 'deploy' phase that's intercepted by the
staging-plugin. Removing it saves additional configuration that
potentially has to be maintained and is not easy to understand.
And because the 'gpg:sign-and-deploy-file' goal is configured directly
in the job as well, the 'ossrh' repository in the distributionManagement
is also effectively unused and can be dropped too.