[Build] Fix and unify invocations of 'update-download-index' job
Update or fix all invocations of a 'update-download-index' job to invoke
only 'Releng/updateIndex', because its definition is under
version-control.
The other previously referenced jobs are either manually
defined/maintained (with equal definition) or don't exist.

This fixes the index-update for the mark-(un)stable and makeVisbile
jobs and allows to delete the manually defined
'eclipse.releng.updateIndex' job.