[Build] Add Jenkins pipeline to improve release preparation automation
This introduces a new Jenkins job named `Prepare Next Development Cycle`
that carries out many of the preparation tasks and aims to further
enhance the release preparation automation. This includes
- Updating the version of the Maven parent, all references to it and the
Eclipse products
- Updating the release version across build scripts
- Updating the previous release version to the current Release-Candidate

These tasks are currently done partly manually and partly automatically,
but through a mixture of the (hidden) Jenkins job to create-milestones
and distributed GH workflows that perform the Maven version updates upon
the creation of new milestones.
Having one central job simplifies to understand the preparation process.
It also ensures that the steps are always executed in the right order.

The script 'getPreviousRelease.sh' is not considered as it seems to be
unused and is intended to be removed soon.