Use buildtools snapshots
With https://github.com/eclipse-platform/eclipse.platform.releng.buildtools/pull/83
latest build of buildtools is deployed to
https://download.eclipse.org/eclipse/updates/buildtools/snapshots/ .
At https://download.eclipse.org/eclipse/updates/buildtools/ there is
composite repo but maintenance of it broke with missing Java 21 on
downloads as scripts run serverside.
This is the simplest and fastest way to put us back in the game.