[Build] Use latest commit of branch specified in .gitmodules file
This simplifies trying out changes in a submodule and their effect on
the entire Eclipse-SDK in advance. Instead of always fetching the latest
commit from the master branch of the specified remote, fetch the latest
commit of the branch specified in '.gitmodules' from each submodule's
remote (also specified in '.gitmodules'). On the CLI this can be
archived by the following command:
'''
git submodule update --remote
'''
But since the Jenkins job is already configured to checkout the
submodules this way, nothing actually has to be done and the
corresponding stage can just be removed.

In order to try out a change with the entire Eclipse-SDK, now one just
has to update the affected submodules in the '.gitmodules' file to
point to the git-repository and the branch providing the change(s).

Before, one either had to push changes to the 'master' branch of the
alternative repository providing a change or had to disable the fetch in
this Jenkins pipeline and record the change's commit in this aggregator
repository along with changing the of the remote in '.gitmodules'.
Of course all of this had to be known.

Additionally add a comment explaining how to archive the just built
product with a Jenkins build.