Add a profile to enable compilation with javac
Sometimes there is a question "does it compile with javac" and doing so
by hand can become cumbersome.

This now adds a new profile "javac" that can be used to enable javac as
a compiler instead of ecj and disable baseline compare/replace because
of javac emits different bytecode by default.