[Build] Enable automated version increments for eclipse.platform.releng
Set up for this repository the workflow for automated version increments
in pull-requests added via
- https://github.com/eclipse-platform/eclipse.platform.releng.aggregator/pull/2352

But only for the projects in the 'eclipse.platform.releng' in order to
avoid building the entire SDK just to check for version increments.
Missing version increments in sub-modules cannot be handled in this
repository anyway.