Warn if workflow for automated version increments cannot modify PR
It it's a constant source of confusion that the commit with version
increments is not pushed to the PR, in case edits by maintainers are
disabled or the PR is created from an organization repository (which
makes edits in general impossible).
This now adds a warning for that case to at least explain contributors
the situation.