[Build] Unify clean-up jobs as Jenkins pipelines and inline scripts
Migrate all clean-up jobs to pure Jenkinsfile based pipeline jobs.

In this process the following enhancements are applied additionally:
- Unify the daily clean-up builds into one job that handles Eclipse and
Equinox builds in parallel.
- Implement the logic for all release/stable and I-build drop removals
upon releases into the job for the cleanup of release artifacts, so that
it doesn't need any input parameters.
- Continuously clean-up Y-build drops, with the policy to keep the last
nine builds (corresponds to three consecutive weeks on the regular build
schedule).
- Introduce the ability to manually mark a as kept-indfinitivly in order
to prevent the clean-up jobs from deleting it.
- Move the update of the eclipse/updates/index.html to redirect to the
update-site of the latest release into the (release) publication job.