[Build] Restore compressed content.xml for platform repository
Include the additional IU 'org.eclipse.equinox.executable' directly in
the platform P2-repository via the defining 'category.xml'. This avoids
the post-processing of the raw 'content.xml' file using an XSL
transformation and consequently ensures the 'content.xml' is only
provided in compressed form as 'content.jar' and 'content.xml.xz'.
Due to the previous post-processing the content.xml was only provided
uncompressed, which increases the load on the download-server, slows
down metadata downloads and also complicates the RelEng process.

Fixes https://github.com/eclipse-platform/eclipse.platform.releng.aggregator/issues/1912

This was originally introduced via
- https://github.com/eclipse-platform/eclipse.platform.releng.aggregator/pull/3187

and reverted due to a regression in I-builds via
- https://github.com/eclipse-platform/eclipse.platform.releng.aggregator/pull/3189

As the cause of that regression is fixed, the change is introduced again
unchanged and
This reverts commit d4d05423bd2f76fa721006d35bf576bd8f833c77.