Convert remaining XHTML documents in platform.doc.isv to HTML
Unlike plain HTML, XHTML also follows the XML rules and is therefore
more restrictive. By mixing both HTML and XHTML documents, this can very
easily lead to a situation where syntax is used that is valid in one,
but not the other document.

With this change, all documents within org.eclipse.platform.doc.isv that
are still using XHTML are converted to HTML.