[RelEng] Update build configuration in release preparation pipeline
and simplify the CRON expressions that define the I/Y-build schedules:
- Drop the RC phase with additional builds because that late in the
dev-cycle many changes are not expected/desired.
If necessary an extra build can still be triggered manually.
- Define the CRON expressions for I-builds so that the schedule
automatically ends with the end of RC2. This avoids the need to change
the CRON-trigger definition manually at the end of the RC phase.