[RelEng] Create and update maintenance branches in release preparation
In the release preparation pipeline create the R4_XX_maintenance branch
for the eclipse.platform.releng.aggregator repo and all submodules.
Furthermore create the commit for the maintenance branch in the
aggregator repo to move the maintained release's I-build to that
maintenance branch and push it as PR against the maintenance branch.