[RelEng] Make read of releaseName more robust in publication pipeline
Don't launch Maven to output the interpolated 'releaseName' property,
but instead read the 'releaseYear' and 'releaseMonth' properties from
the raw (eclipse-platform-parent/pom.xml) file.
This avoids problems if at the time of the release (candidate) build a
snapshot version was used, e.g. of Tycho, that's not available anymore
at the time of the publication (because e.g. that Tycho version was
released in the meantime).