[RelEng] Consider next java release date for Y-build schedule
Explicitly specify a limited cron based schedule for Y-builds too that
is either aligned with the Eclipse release or a Java-release.
If no Java release happens soon after the Eclipse release the schedule
ends two days before RC2 (similar to I-builds).
If a Java release happens soon after the Eclipse release (and a JDT
patch is provided after the Eclipse release), the Y-build schedule is
extended until the java release date and the schedule and the Y-build
cadence is increased in the 'RC-phase'.