[RelEng] Split stream, branch and schedule configuration for I/Y builds
Split the configuration of built branches per stream and specify it
individually per build type (I- and Y-build). Additionally specify
individually per build type and stream the cron trigger schedule of the
corresponding build.
This allows to configure all managed builds individually.

When preparing a new release, add new and move old stream as before. But
preserve the old build's schedule and don't overwrite especially the
Y-build schedule with the new streams schedule even if a Java release is
imminent (this wrongly happened before).
For Y builds, also add the new stream immediately, but disable it, when
a java release is imminent and the the new BETA-Java branch is not
created. The old stream's Y-build can then continue until the Java
release to feed the P-build and the new build can be activated once
ready.