[I/Y-Build] Unify build configurations into buildConfigurations.json
The buildConfigurations.json becomes the central and only source for
configurations of all I/Y-builds and their associated tests.
It's read by the seed-job to create the corresponding Jenkins jobs and
during I/Y-builds executions.

This avoids a definition of test configurations split over multiple
files and avoids continuous changes in the build pipeline file due to
regular version, Java or infrastructure updates.