[Build] Read parameters from Jenkins-files and create jobs dynamically
When creating a pipeline-job the Jenkins JobDSL, by default, doesn't
consider the parameters defined within the pipeline-job's Jenkinsfile.
Therefore parameters defined only within the Jenkinsfile are only added
to the job's definition after the pipeline was executed first (and
updated on subsequent executions, if changes are applied).
In order to make parameters available even on the first run, instead
they have to be defined in the groovy-file that uses the Jenkins Job-DSL
to define the pipeline-job.
This is an long standing and known, yet unsolved issue in Jenkins:
- https://issues.jenkins.io/browse/JENKINS-41929

The current solution to define parameters in the job-definition has the
drawback that an important part of the pipeline is defined in another
file, which makes maintenance more difficult.

This implements an enhanced work-around, allowing to define the
parameters within a pipeline's Jenkinsfile again:
During the definition of a pipeline-job, it's Jenkinsfile file is parsed
and its AST is searched for parameter definitions. All encountered
parameters are translated dynamically into corresponding parameters of
the job definition.

With this new possibility all parameters have their definition moved
into their pipeline's Jenkinsfile.

Furthermore this adds a JenkinsFile for the seedJob of the RelEng JIPP,
which was previously manually defined as a free-style job.

That seedJob pipeline also creates jobs dynamically from all pipeline
files in the corresponding folders. It's therefore sufficient to just
add a JenkinsFile and not necessary anymore to explicitly add a
corresponding pipeline job definition anywhere.

This also ensures a unified naming schema of all jobs that corresponds
to the name of the pipeline's Jenkinsfile.