[Build] Streamline and speed-up test-result collection
Instead of downloading the entire drop of a build to the build-agent,
just download the required test-result XML files of other test-runs of
the same build.
Also avoid the use of raw.githubusercontent URLs as their use is
rate-limited.
Furthermore remove unused elements in 'publish.xml' and
'eclipse/helper.xml' after the performance-test infrastructure was
removed.