[RelEng] Fix access to empty DRY_RUN_PREFIX in promotion pipeline
If an environment variable is defined with an empty value (as it's the
case for 'DRY_RUN_PREFIX' if a build is not a dry-run) it's effectively
discarded, not passed to scripts and is not available as environment
variable binding respectively in the special 'env' map/variable.

Implement corresponding special handling for that case of a non-dry-run.