[I/Y-Build] Derive git branch and URL from job definition when cloning
This reduces hard-coded branch names in the pipeline file.
It also avoids the need to update them, e.g. on maintenance branches
respectively simplifies testing changes. As one can either modify the
git URL and branch-name in one place or they are derived if a modified
copy of the build jobs is used, pointing to a fork of this repository
and/or another branch.

This also reduces the changes necessary during the preparation of a new
development cycle.

Furthermore this change minimizes the time the SSH-agent is running and
unifies git-pushing.