[Build] Specify timeout for test execution after agent was accessed
The top-level timeout specified in the global options of a Jenkins job
is effective from the time the job starts its execution (waiting for the
'main' agents is excluded here). On the contrary timeouts defined on
stage level include the time waiting for the stage's agent to become
available.
In some cases (currently only for Mac) the same physical machine is used
to execute multiple test-configurations. Consequently some
configurations have to wait a significant amount of time for the
executing agent to become available.

Therefore now timeouts are defined:
An inner one effective only for the actual test execution, sufficient to
execute one configuration.
And an outer, global one, effective almost from when the job is
triggered and with a larger value to cover the execution of multiple
configurations.
The global one mainly exists in order to prevent pending tests to pile
up in case the infrastructure is flawed.

Follow-up on
- https://github.com/eclipse-platform/eclipse.platform.releng.aggregator/pull/3416