[I/Y-Build] Capture build logs via Jenkins stage view API
instead of capturing them through the shell using tee.
This simplifies to capture the logs and allows capturing the logs of
stages that are (completely) defined within the Jenkins pipeline and not
only those that are defined in scripts called from the pipeline.

Additionally make references to CSS style-sheets absolute.