[Build] Materialize products from local repository to add pgp signatures
By default products are materialized from the target-platform, but the
pgp-sigatures produced in a previous execution of the tycho-pgp-plugin
are only added to the local repository of the product (in it's
'target/repository' directory).
In order to make sure that the produced pgp-sigatures are included into
the products (i.e. their embedded repositories) the later repository has
to be the source of the materialization.

Follow-up on
- https://github.com/eclipse-platform/eclipse.platform.releng.aggregator/pull/3480