[Build] Add back sources to Eclipse-SDK product
The sources of all contained bundles were accidentally removed from the
eclipse-SDK product in
- https://github.com/eclipse-platform/eclipse.platform.releng.aggregator/pull/3549

That is because the p2-repository assembled for the eclipse-SDK product
didn't contain sources before, as that repository is not used further.
But now that the product is materialized from that p2-repository (since
that mentioned change), sources went missing in the product itself too.

Consequently the property 'repository.includeAllSources' must now be
true if 'product.installSources' is true. Otherwise the latter has no
effect. If no sources are available in the source-repository, of course
Tycho cannot install them.

Additionally remove dependencies again from eclipse-junit-tests product.
They were added unintentionally in the mentioned change, too.

Fixes
- https://github.com/eclipse-jdt/eclipse.jdt.core/issues/4684